+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # Change log
+<hr>
+
+## [1.3.19] 2023-09-05
+
+### Added
+
+- API Core - Operator Outbound Scan parcel using data below:
+- API Core - Operator check order {string} have the following Tags:
+- API Core - Operator update parcel size:
+- API Core - set system parameter:
+- API Core - verify driver's total cod:
+- API Route - delete routes:
+- DB Core - Operator search {string} Orders with {string} Status and {string} Granular Status
+- DB Core - soft delete route {value}
+- DB Core - verify total COD for driver:
+- DB Route - get route_logs record for driver id {string}
 
 <hr>